Further breaking things into smaller components
