+++ frontend/src/components/admin/MLDashboard/models.js
@@ -4,7 +4,7 @@ import ReactTable from 'react-table';
 import {BorderDiv, TableRow} from './util/styledDivs';
 import GameChangerAPI from '../../api/gameChanger-service-api';
 import GCPrimaryButton from "../../common/GCButton";
-import styles from '../GCAdminStyles';
+import {styles} from '../GCAdminStyles';
 import "react-table/react-table.css";
 import './index.scss';
